Move snarkit2 test folder into test and add test cases for both zero
