+++ beacon-light-client/circom/test/subtract/circuit.circom
@@ -1,0 +1,5 @@
+pragma circom 2.0.3;
+
+include "../../../../vendor/circom-pairing/circuits/curve.circom";
+
+component main = EllipticCurveSubtract(55, 7, 0, 4, [35747322042231467, 36025922209447795, 1084959616957103, 7925923977987733, 16551456537884751, 23443114579904617, 1829881462546425]);
