+++ beacon-light-client/circom/test/subtract/data/both_zero/input.json
@@ -1,0 +1,1 @@
+{"a":[["16589478066046651","22658679592837110","35004527604248919","16789302793630161","7530538873701625","32234187716135413","1684953952445941"],["11860609209853921","4091579406338073","12578493816062195","13088963032438982","24961455755233629","8501508834176643","612415636564648"]],"b":[["16589478066046651","22658679592837110","35004527604248919","16789302793630161","7530538873701625","32234187716135413","1684953952445941"],["11860609209853921","4091579406338073","12578493816062195","13088963032438982","24961455755233629","8501508834176643","612415636564648"]],"aIsInfinity":1,"bIsInfinity":1}
