+++ beacon-light-client/circom/test/subtract/data/case0/input.json
