+++ beacon-light-client/circom/test/subtract/data/case1/input.json
