+++ beacon-light-client/circom/test/subtract/data/points_eqaul/input.json
