+++ beacon-light-client/circom/test/subtract/data/subtract_from_zero/input.json
