+++ beacon-light-client/circom/test/subtract/data/subtract_zero/input.json
