+++ beacon-light-client/circom/test/subtract/data/subtract_zero/output.json
